[FEAT] 팀 여부에 따라 보여지는 컴포넌트 달라지는 컴포넌트 #NADA-60
