+++ src/Badge/containers/register/PeopleForm.js
@@ -1,25 +1,10 @@
 import { TitleBox } from '../../../styles/Survey';
-import { useDispatch, useSelector } from 'react-redux';
-import IssueListInputItem from '../../components/IssueListInputItem';
-import { ButtonContainer } from './ActivityForm';
-import {
-  Act,
-  Cancel,
-} from '../../../Modal/components/usedInModal/ModalButtonDiv';
-import { addList } from '../../modules/redux/badge';
+import { useSelector } from 'react-redux';
+
+import PeopleInfoInput from '../../components/PeopleInfoInput';
 
 const PeopleForm = ({ onSubmit }) => {
-  const issueList = useSelector(({ badge }) => badge.issueList);
-  const dispatch = useDispatch();
-
-  const onClick = () => {
-    dispatch(
-      addList({
-        type: 'issueList',
-        value: { name: '', role: '', birth: '', email: '', number: '' },
-      }),
-    );
-  };
+  const teams = useSelector(({ badge }) => badge.teams);
 
   return (
     <>
@@ -29,29 +14,8 @@ const PeopleForm = ({ onSubmit }) => {
         정보를 입력해 주세요
       </TitleBox>
 
-      <div
-        style={{
-          overflowY: 'scroll',
-          display: 'flex',
-          flexDirection: 'column',
-          gap: '24px',
-        }}
-      >
-        {issueList.map((el, index) => (
-          <IssueListInputItem key={index} content={el} index={index} />
-        ))}
-      </div>
-
-      <ButtonContainer>
-        <Cancel onClick={() => onClick()}>
-          <div className="text">발급자 추가하기</div>
-        </Cancel>
-        <Act onClick={onSubmit}>
-          <div className="text" style={{ color: 'white' }}>
-            다음
-          </div>
-        </Act>
-      </ButtonContainer>
+      {/* 팀 추가를 하지 않았을때, */}
+      {teams.length === 0 && <PeopleInfoInput onSubmit={onSubmit} />}
     </>
   );
 };
